fix(ledger): reject narrow Dijkstra redeemer indices

+++ ledger/common/script/purpose.go
@@ -328,7 +328,7 @@ func scriptPurposeBuilder(
 	) (ScriptPurpose, error) {
 		switch redeemerKey.Tag {
 		case lcommon.RedeemerTagSpend:
-			if int(redeemerKey.Index) >= len(inputs) {
+			if uint64(redeemerKey.Index) >= uint64(len(inputs)) {
 				return nil, UnmatchedRedeemerError{RedeemerKey: redeemerKey}
 			}
 			var datum data.PlutusData
@@ -363,7 +363,7 @@ func scriptPurposeBuilder(
 			}, nil
 		case lcommon.RedeemerTagMint:
 			mintPolicies := mint.Policies()
-			if int(redeemerKey.Index) >= len(mintPolicies) {
+			if uint64(redeemerKey.Index) >= uint64(len(mintPolicies)) {
 				return nil, UnmatchedRedeemerError{RedeemerKey: redeemerKey}
 			}
 			slices.SortFunc(
@@ -374,7 +374,7 @@ func scriptPurposeBuilder(
 				PolicyId: mintPolicies[redeemerKey.Index],
 			}, nil
 		case lcommon.RedeemerTagCert:
-			if int(redeemerKey.Index) >= len(certificates) {
+			if uint64(redeemerKey.Index) >= uint64(len(certificates)) {
 				return nil, UnmatchedRedeemerError{RedeemerKey: redeemerKey}
 			}
 			return ScriptPurposeCertifying{
@@ -382,7 +382,7 @@ func scriptPurposeBuilder(
 				Certificate: certificates[redeemerKey.Index],
 			}, nil
 		case lcommon.RedeemerTagReward:
-			if int(redeemerKey.Index) >= len(withdrawals) {
+			if uint64(redeemerKey.Index) >= uint64(len(withdrawals)) {
 				return nil, UnmatchedRedeemerError{RedeemerKey: redeemerKey}
 			}
 			return ScriptPurposeRewarding{
@@ -392,14 +392,14 @@ func scriptPurposeBuilder(
 				},
 			}, nil
 		case lcommon.RedeemerTagVoting:
-			if int(redeemerKey.Index) >= len(votes) {
+			if uint64(redeemerKey.Index) >= uint64(len(votes)) {
 				return nil, UnmatchedRedeemerError{RedeemerKey: redeemerKey}
 			}
 			return ScriptPurposeVoting{
 				Voter: *votes[redeemerKey.Index].Key,
 			}, nil
 		case lcommon.RedeemerTagProposing:
-			if int(redeemerKey.Index) >= len(proposalProcedures) {
+			if uint64(redeemerKey.Index) >= uint64(len(proposalProcedures)) {
 				return nil, UnmatchedRedeemerError{RedeemerKey: redeemerKey}
 			}
 			return ScriptPurposeProposing{
@@ -433,7 +433,7 @@ func BuildScriptPurpose(
 ) (ScriptPurpose, error) {
 	switch redeemerKey.Tag {
 	case lcommon.RedeemerTagSpend:
-		if int(redeemerKey.Index) >= len(inputs) {
+		if uint64(redeemerKey.Index) >= uint64(len(inputs)) {
 			return nil, UnmatchedRedeemerError{RedeemerKey: redeemerKey}
 		}
 		tmpInput := inputs[redeemerKey.Index]
@@ -460,7 +460,7 @@ func BuildScriptPurpose(
 		}, nil
 	case lcommon.RedeemerTagMint:
 		policies := mint.Policies()
-		if int(redeemerKey.Index) >= len(policies) {
+		if uint64(redeemerKey.Index) >= uint64(len(policies)) {
 			return nil, UnmatchedRedeemerError{RedeemerKey: redeemerKey}
 		}
 		slices.SortFunc(
@@ -471,7 +471,7 @@ func BuildScriptPurpose(
 			PolicyId: policies[redeemerKey.Index],
 		}, nil
 	case lcommon.RedeemerTagCert:
-		if int(redeemerKey.Index) >= len(certificates) {
+		if uint64(redeemerKey.Index) >= uint64(len(certificates)) {
 			return nil, UnmatchedRedeemerError{RedeemerKey: redeemerKey}
 		}
 		return ScriptPurposeCertifying{
@@ -493,7 +493,7 @@ func BuildScriptPurpose(
 			}
 			return bytes.Compare(aBytes, bBytes)
 		})
-		if int(redeemerKey.Index) >= len(sortedAddrs) {
+		if uint64(redeemerKey.Index) >= uint64(len(sortedAddrs)) {
 			return nil, UnmatchedRedeemerError{RedeemerKey: redeemerKey}
 		}
 		addr := sortedAddrs[redeemerKey.Index]
@@ -536,14 +536,14 @@ func BuildScriptPurpose(
 			}
 			return 1
 		})
-		if int(redeemerKey.Index) >= len(sortedVoters) {
+		if uint64(redeemerKey.Index) >= uint64(len(sortedVoters)) {
 			return nil, UnmatchedRedeemerError{RedeemerKey: redeemerKey}
 		}
 		return ScriptPurposeVoting{
 			Voter: *sortedVoters[redeemerKey.Index],
 		}, nil
 	case lcommon.RedeemerTagProposing:
-		if int(redeemerKey.Index) >= len(proposalProcedures) {
+		if uint64(redeemerKey.Index) >= uint64(len(proposalProcedures)) {
 			return nil, UnmatchedRedeemerError{RedeemerKey: redeemerKey}
 		}
 		return ScriptPurposeProposing{
